refactor(text): adjust the implementation of Themed
