+++ src/components/data/Text/styles.ts
@@ -2,26 +2,31 @@ import styled from "styled-components";
 
 import { inube } from "@shared/tokens";
 import { ITextProps } from ".";
+import { Themed } from "@src/shared/Types/types";
+
+interface IStyledTextProps extends ITextProps {
+  theme: Themed;
+}
 
 const StyledText = styled.p`
-  font-family: ${({ type, size, theme }: ITextProps) =>
+  font-family: ${({ type, size, theme }: IStyledTextProps) =>
     theme?.typography?.[type!]?.[size!]?.font ||
     inube.typography[type!][size!].font};
-  line-height: ${({ type, size }: ITextProps) =>
+  line-height: ${({ type, size }: IStyledTextProps) =>
     inube.typography[type!][size!].lineHeight};
-  font-size: ${({ type, size, theme }: ITextProps) =>
+  font-size: ${({ type, size, theme }: IStyledTextProps) =>
     theme?.typography?.[type!]?.[size!].size ||
     inube.typography[type!][size!].size};
-  letter-spacing: ${({ type, size, theme }: ITextProps) =>
+  letter-spacing: ${({ type, size, theme }: IStyledTextProps) =>
     theme?.typography?.[type!]?.[size!]?.tracking ||
     inube.typography[type!][size!].tracking};
-  font-weight: ${({ type, size, theme }: ITextProps) =>
+  font-weight: ${({ type, size, theme }: IStyledTextProps) =>
     theme?.typography?.[type!]?.[size!]?.weight ||
     inube.typography[type!][size!].weight};
-  margin: ${({ margin }: ITextProps) => margin};
-  padding: ${({ padding }: ITextProps) => padding};
-  text-align: ${({ textAlign }: ITextProps) => textAlign};
-  color: ${({ appearance, disabled, parentHover, theme }: ITextProps) => {
+  margin: ${({ margin }: IStyledTextProps) => margin};
+  padding: ${({ padding }: IStyledTextProps) => padding};
+  text-align: ${({ textAlign }: IStyledTextProps) => textAlign};
+  color: ${({ appearance, disabled, parentHover, theme }: IStyledTextProps) => {
     if (disabled) {
       return (
         theme?.color?.text?.[appearance!]?.disabled ||
@@ -40,14 +45,14 @@ const StyledText = styled.p`
     );
   }};
 
-  white-space: ${({ ellipsis }: ITextProps) => ellipsis && "nowrap"};
-  overflow: ${({ ellipsis }: ITextProps) => ellipsis && "hidden"};
-  text-overflow: ${({ ellipsis }: ITextProps) => ellipsis && "ellipsis"};
-  cursor: ${({ cursorHover, parentHover }: ITextProps) =>
+  white-space: ${({ ellipsis }: IStyledTextProps) => ellipsis && "nowrap"};
+  overflow: ${({ ellipsis }: IStyledTextProps) => ellipsis && "hidden"};
+  text-overflow: ${({ ellipsis }: IStyledTextProps) => ellipsis && "ellipsis"};
+  cursor: ${({ cursorHover, parentHover }: IStyledTextProps) =>
     (cursorHover || parentHover) && "pointer"};
 
   &:hover {
-    color: ${({ appearance, disabled, cursorHover, theme }: ITextProps) =>
+    color: ${({ appearance, disabled, cursorHover, theme }: IStyledTextProps) =>
       !disabled &&
       cursorHover &&
       (theme?.color?.text?.[appearance!]?.hover ||
